Use a uint16_t for gpio_base
16-bit value makes more sense in this case (16-bit IO address space).
The actual GPIOBASE register is 32-bits wide, but the upper 16 are
reserved.



git-svn-id: https://mosys.googlecode.com/svn/trunk@457 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e

+++ include/drivers/intel/ich_generic.h
@@ -89,7 +89,7 @@ int ich_set_bbs(struct platform_intf *intf, int bbs);
  * returns 0 to indicate success
  * returns <0 on read failure
  */
-extern int ich_get_gpio_base(struct platform_intf *intf, uint32_t *val);
+extern int ich_get_gpio_base(struct platform_intf *intf, uint16_t *val);
 
 /*
  * ich_read_gpio  - read GPIO status
